requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
absl-py==2.1.0
    # via
    #   keras
    #   tensorboard
    #   tensorflow
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
astunparse==1.6.3
    # via tensorflow
certifi==2024.6.2
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   typer
    #   uvicorn
dnspython==2.6.1
    # via email-validator
dynaconf==3.2.5
    # via -r requirements.in
email-validator==2.1.1
    # via fastapi
exceptiongroup==1.2.1
    # via anyio
fastapi==0.111.0
    # via -r requirements.in
fastapi-cli==0.0.4
    # via fastapi
flatbuffers==24.3.25
    # via tensorflow
gast==0.5.4
    # via tensorflow
google-pasta==0.2.0
    # via tensorflow
grpcio==1.64.1
    # via
    #   tensorboard
    #   tensorflow
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
h5py==3.11.0
    # via
    #   keras
    #   tensorflow
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via fastapi
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jinja2==3.1.4
    # via fastapi
keras==3.3.3
    # via tensorflow
libclang==18.1.1
    # via tensorflow
markdown==3.6
    # via tensorboard
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
ml-dtypes==0.3.2
    # via
    #   keras
    #   tensorflow
namex==0.0.8
    # via keras
numpy==1.26.4
    # via
    #   h5py
    #   keras
    #   ml-dtypes
    #   opt-einsum
    #   tensorboard
    #   tensorflow
opt-einsum==3.3.0
    # via tensorflow
optree==0.11.0
    # via keras
orjson==3.10.3
    # via fastapi
packaging==24.0
    # via tensorflow
pillow==10.3.0
    # via -r requirements.in
protobuf==4.25.3
    # via
    #   tensorboard
    #   tensorflow
pydantic==2.7.3
    # via
    #   fastapi
    #   pydantic-settings
pydantic-core==2.18.4
    # via pydantic
pydantic-settings==2.3.1
    # via -r requirements.in
pygments==2.18.0
    # via rich
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.9
    # via
    #   -r requirements.in
    #   fastapi
pyyaml==6.0.1
    # via uvicorn
requests==2.32.3
    # via tensorflow
rich==13.7.1
    # via
    #   keras
    #   typer
shellingham==1.5.4
    # via typer
six==1.16.0
    # via
    #   astunparse
    #   google-pasta
    #   tensorboard
    #   tensorflow
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
starlette==0.37.2
    # via fastapi
tensorboard==2.16.2
    # via tensorflow
tensorboard-data-server==0.7.2
    # via tensorboard
tensorflow==2.16.1
    # via -r requirements.in
tensorflow-io-gcs-filesystem==0.37.0
    # via tensorflow
termcolor==2.4.0
    # via tensorflow
typer==0.12.3
    # via
    #   -r requirements.in
    #   fastapi-cli
typing-extensions==4.12.1
    # via
    #   anyio
    #   fastapi
    #   optree
    #   pydantic
    #   pydantic-core
    #   tensorflow
    #   typer
    #   uvicorn
ujson==5.10.0
    # via fastapi
urllib3==2.2.1
    # via requests
uvicorn[standard]==0.30.1
    # via
    #   -r requirements.in
    #   fastapi
uvloop==0.19.0
    # via uvicorn
watchfiles==0.22.0
    # via uvicorn
websockets==12.0
    # via uvicorn
werkzeug==3.0.3
    # via tensorboard
wheel==0.43.0
    # via astunparse
wrapt==1.16.0
    # via tensorflow

# The following packages are considered to be unsafe in a requirements file:
# setuptools
